feat!: migrate to TanStack Start BETA directory structure
BREAKING CHANGE: Complete migration from app/ to src/ directory structure

- Remove old app/ directory structure and scattered files
- Add new src/ directory structure:
  - src/auth-schema.ts (Better Auth schemas)
  - src/components/ (UI components and layouts)
  - src/database/ (database schemas and configuration)
  - src/hooks/ (custom React hooks)
  - src/lib/ (utilities and auth configuration)
  - src/providers.tsx (React providers)
  - src/router.tsx (TanStack Router configuration)
  - src/routes/ (file-based routing structure)
  - src/styles/ (global styles)
  - src/types/ (TypeScript type definitions)
  - src/logo.svg

- Update .gitignore to include migrations directory
- Format components.json with consistent indentation

This follows the TanStack Start BETA migration guide for Vite-native directory structure.
