--- app/api.ts
@@ -1,3 +1,0 @@
-import { createStartAPIHandler, defaultAPIFileRouteHandler } from "@tanstack/react-start/api"
-
-export default createStartAPIHandler(defaultAPIFileRouteHandler)
