--- app/routes/api/auth/$.ts
@@ -1,7 +1,0 @@
-import { auth } from "@/lib/auth"
-import { createAPIFileRoute } from "@tanstack/react-start/api"
-
-export const APIRoute = createAPIFileRoute("/api/auth/$")({
-    GET: ({ request }) => auth.handler(request),
-    POST: ({ request }) => auth.handler(request)
-})
