--- app/ssr.tsx
@@ -1,9 +1,0 @@
-import { getRouterManifest } from "@tanstack/react-start/router-manifest"
-import { createStartHandler, defaultStreamHandler } from "@tanstack/react-start/server"
-
-import { createRouter } from "./router"
-
-export default createStartHandler({
-    createRouter,
-    getRouterManifest
-})(defaultStreamHandler)
