+++ src/routes/__root.tsx
@@ -1,0 +1,77 @@
+import {
+  HeadContent,
+  Outlet,
+  Scripts,
+  createRootRoute,
+} from "@tanstack/react-router";
+import type { ReactNode } from "react";
+
+import { Header } from "@/components/header";
+import { Providers } from "../providers";
+
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+import appCss from "@/styles/globals.css?url";
+
+export const Route = createRootRoute({
+  head: () => ({
+    meta: [
+      {
+        charSet: "utf-8",
+      },
+      {
+        name: "viewport",
+        content: "width=device-width, initial-scale=1",
+      },
+      {
+        title: "Better Auth TanStack Starter",
+      },
+    ],
+    links: [{ rel: "stylesheet", href: appCss }],
+  }),
+  component: RootComponent,
+});
+
+function RootComponent() {
+  return (
+    <RootDocument>
+      <Outlet />
+    </RootDocument>
+  );
+}
+
+function RootDocument({ children }: Readonly<{ children: ReactNode }>) {
+  return (
+    <html lang="en" suppressHydrationWarning>
+      <head>
+        <meta
+          name="viewport"
+          content="initial-scale=1, viewport-fit=cover, width=device-width"
+        />
+        <meta
+          name="theme-color"
+          media="(prefers-color-scheme: light)"
+          content="oklch(1 0 0)"
+        />
+        <meta
+          name="theme-color"
+          media="(prefers-color-scheme: dark)"
+          content="oklch(0.145 0 0)"
+        />
+
+        <HeadContent />
+      </head>
+
+      <body>
+        <Providers>
+          <div className="flex min-h-svh flex-col">
+            <Header />
+
+            {children}
+          </div>
+        </Providers>
+
+        <Scripts />
+      </body>
+    </html>
+  );
+}
