+++ src/routes/auth/$pathname.tsx
