+++ src/types/css-url.d.ts
@@ -1,0 +1,4 @@
+declare module "*.css?url" {
+  const url: string;
+  export default url;
+}
